fix: extract domain logic into internal/allocation (phase 5)
Move buildAllocation, computeHostnames, computeProtocols, computeEnvVars,
buildTemplateVars, and resolveComputedFromAlloc from cmd/up.go into a new
internal/allocation package. These are pure domain functions — they transform
config + ports into allocations and template maps with no CLI dependencies.

up.go: 534 → 214 lines. New package: internal/allocation (122 lines).
