+++ cmd/status.go
@@ -7,6 +7,7 @@ import (
 	"slices"
 
 	"charm.land/lipgloss/v2"
+	"github.com/outport-app/outport/internal/allocation"
 	"github.com/outport-app/outport/internal/certmanager"
 	"github.com/outport-app/outport/internal/config"
 	"github.com/outport-app/outport/internal/portcheck"
@@ -146,7 +147,7 @@ func printStatusJSON(cmd *cobra.Command, reg *registry.Registry, projects map[st
 
 		var computed map[string]computedJSON
 		if cfg != nil && statusComputedFlag {
-			computed = buildComputedMap(cfg.Computed, resolveComputedFromAlloc(cfg, instanceName, alloc.Ports, alloc.Hostnames, httpsEnabled, nil))
+			computed = buildComputedMap(cfg.Computed, allocation.ResolveComputed(cfg, instanceName, alloc.Ports, alloc.Hostnames, httpsEnabled, nil))
 		}
 
 		entries = append(entries, statusEntryJSON{
@@ -199,7 +200,7 @@ func printStatusStyled(cmd *cobra.Command, reg *registry.Registry, projects map[
 		}
 
 		if cfg != nil && statusComputedFlag {
-			if resolved := resolveComputedFromAlloc(cfg, instanceName, alloc.Ports, alloc.Hostnames, httpsEnabled, nil); len(resolved) > 0 {
+			if resolved := allocation.ResolveComputed(cfg, instanceName, alloc.Ports, alloc.Hostnames, httpsEnabled, nil); len(resolved) > 0 {
 				printComputedValues(w, resolved)
 			}
 		}
